// src/app/api/swagger/models/api-trade-ext-customer-info-excel-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface ApiTradeExtCustomerInfoExcelDto {
  BankName?: string | null;
  BicBank?: string | null;
  CheckingAccount?: string | null;
  ContractEndDate?: string | null;
  ContractNumber?: string | null;
  ContractProlongationNoChanges?: string | null;
  ContractResponsibleEmail?: string | null;
  ContractResponsibleName?: string | null;
  ContractResponsiblePatronimic?: string | null;
  ContractResponsiblePhoneNumber?: string | null;
  ContractResponsiblePosition?: string | null;
  ContractResponsibleSurname?: string | null;
  ContractStartDate?: string | null;
  EdmOperator?: string | null;
  FaxNumber?: string | null;
  SignatoryAuthorityDocument?: string | null;
  SignatoryName?: string | null;
  SignatoryPatronimic?: string | null;
  SignatoryPosition?: string | null;
  SignatorySurname?: string | null;
}